CORS headers and tests writen, assuming CORS is always the case

+++ spec/features/mock_cors_spec.rb
@@ -9,7 +9,7 @@ describe Pact::Consumer::MockService do
   CORS_LOG_PATH = File.join File.dirname(__FILE__), 'log', 'mock_cors_spec.log'
 
   before :all do
-    FileUtils.rm CORS_LOG_PATH
+    FileUtils.rm CORS_LOG_PATH if File.exist?(CORS_LOG_PATH)
   end
 
   let(:log_file) { File.open CORS_LOG_PATH, 'a' }
@@ -27,7 +27,7 @@ describe Pact::Consumer::MockService do
         method: :post,
         path: '/alligators/new',
         headers: { 'Accept' => 'application/json' },
-        body: { id: 123, name: 'Mary'}
+        body: { id: 123, name: 'Mary'}.to_json
       },
       response: {
         status: 200,
@@ -37,31 +37,30 @@ describe Pact::Consumer::MockService do
     }.to_json
   end
 
-  context "when a response has been mocked" do
-    context "when the actual request is an OPTIONS prior to the actual request" do
-      it "returns a 200 to allow the request" do | example |
+  context "when in a cross domain environment (CORS)" do
+    context "when a request has been mocked" do
+      it "answers the OPTIONS request, and then appropiately mocks the actual request" do | example |
         # Clear interactions - this would typically be done in a before hook
         delete "/interactions?example_description=#{CGI::escape(example.full_description)}", nil, admin_headers
 
         # Set up expected interaction - this would be done by the Pact DSL
         post "/interactions", expected_interaction, admin_headers
 
-        # Invoke the actual request - this would be done by the class under test
+        # Make the preflight request - this one will not have been created by the user
         options '/alligators/new', nil, { 'HTTP_Access_Control_Request_Headers' => 'x-pact-mock-service, application/json' }
 
-        # Ensure that the response we get back was the one we expected
-        # A test using pact would normally check the object returned from the class under test
-        # eg. expect(client.alligators).to eq [Alligator.new(name: 'Mary')]
+        # Ensure it allows the browser to actually make the request
         expect(last_response.status).to eq 200
         expect(last_response.header).to include "Access-Control-Allow-Origin"=>"*"
         expect(last_response.header).to include "Access-Control-Allow-Headers"=>"x-pact-mock-service, application/json"
         expect(last_response.header).to include "Access-Control-Allow-Methods"=>"DELETE, POST, GET, HEAD, PUT, TRACE, CONNECT"
 
+        # Make the request
+        post "/alligators/new", { id: 123, name: 'Mary'}.to_json , { 'HTTP_ACCEPT' => 'application/json' }
 
-        puts last_response.headers.inspect
-        skip
-
-        # Needs to check, when you make the actual post, that the CORS is actually set
+        # Ensure that the response we get back was the one we expected
+        # and includes the CORS header
+        expect(last_response.header).to include "Access-Control-Allow-Origin"=>"*"
         expect(last_response.headers['Content-Type']).to eq 'application/json'
         expect(JSON.parse(last_response.body)).to eq([{ 'name' => 'Mary' }])
 
@@ -73,16 +72,24 @@ describe Pact::Consumer::MockService do
     end
 
     context "when the actual request does not match the expected request" do
-      it "returns an error response" do | example |
-        skip
+      it "successfully answers the OPTIONS request, and then returns an error message on the request per say" do | example |
         # Clear interactions - this would typically be done in a before hook
         delete "/interactions?example_description=#{CGI::escape(example.full_description)}", nil, admin_headers
 
         # Set up expected interaction - this would be done by the Pact DSL
         post "/interactions", expected_interaction, admin_headers
 
-        # Invoke the actual request - this would be done by the class under test
-        get "/alligators", nil, { 'HTTP_ACCEPT' => 'application/xml' }
+        # Make the preflight request - this one will not have been created by the user
+        options '/alligators/new', nil, { 'HTTP_Access_Control_Request_Headers' => 'x-pact-mock-service, application/json' }
+
+        # Ensure it allows the browser to actually make the request
+        expect(last_response.status).to eq 200
+        expect(last_response.header).to include "Access-Control-Allow-Origin"=>"*"
+        expect(last_response.header).to include "Access-Control-Allow-Headers"=>"x-pact-mock-service, application/json"
+        expect(last_response.header).to include "Access-Control-Allow-Methods"=>"DELETE, POST, GET, HEAD, PUT, TRACE, CONNECT"
+
+        # Make the request
+        post "/alligators/new", { id: 124, name: 'John'}.to_json , { 'HTTP_ACCEPT' => 'application/json' }
 
         # A 500 is returned as the headers don't match
         # An actual test should fail at this point as the class under test would probably raise an exception
